Move applicant components to applicant folder
