+++ src/components/Applicant/ApplicantDetail/component.tsx
