+++ src/components/Applicant/ApplicantDetail/container.tsx
