+++ src/components/Applicant/ApplicantDetail/index.tsx
