+++ src/components/Applicant/ApplicantDetail/interface.tsx
