+++ src/components/Applicant/ApplicantDetail/styles.module.scss
