+++ src/components/Applicant/ApplicantDetailEditable/interface.tsx
