+++ src/components/Applicant/ApplicantDetailEditable/styles.module.scss
